p3_bookings
Created a scrapper for booking.com to extract the hotels data and save it to csv file.

Usage: We have to go to the website and applying the filters about hotels that we want. After that, once the page loaded, we have to paste the link in scrapper and also give the csv file_name that we want to create for saving the data.

Function: The scrapper will extract all the hotels data. It will extract the hotel name, location, price, rating, score, reviews and url of each hotel.

Stack used: requests, BeautifulSoup, csv